+++ src/c++/lib/starling_common/starling_read_align.cpp
@@ -821,7 +821,6 @@ getCurIndelHaplotypeIds(
     const unsigned sampleCount(curIndelHaplotypeIds.size());
 
     ActiveRegionId curIndelActiveRegionId(curIndelData.activeRegionId);
-    curIndelHaplotypeIds.clear();
     if (curIndelActiveRegionId < 0) return;
 
     // Get haplotype IDs of current indel in all samples
@@ -869,7 +868,6 @@ candidate_alignment_search(
     mca_warnings& warn,
     starling_align_indel_status indel_status_map,
     HaplotypeStatusMap haplotypeStatusMap,
-    const bool containsIndelWithNoReadSupport,
     std::vector<IndelKey> indel_order,
     const unsigned depth,
     const unsigned indelToggleDepth,
@@ -994,12 +992,18 @@ candidate_alignment_search(
     const IndelKey& curIndel(indel_order[depth]);
 
     bool isCurIndelConflicting(false);
+    bool containsIndelNotDiscoveredFromReads(false);
     for (unsigned i(0); i<depth; ++i)
     {
         const IndelKey& indelKey(indel_order[i]);
 
         if (!(indel_status_map[indelKey].is_present)) continue;
         if (is_indel_conflict(indelKey, curIndel)) isCurIndelConflicting = true;
+        if ((! containsIndelNotDiscoveredFromReads) &&
+            (indelBuffer.getIndelDataPtr(indelKey)->status.notDiscoveredFromReads))
+        {
+            containsIndelNotDiscoveredFromReads = true;
+        }
     }
 
     const unsigned sampleCount(opt.getSampleCount());
@@ -1018,8 +1022,8 @@ candidate_alignment_search(
         }
     }
 
-    // true if current indel is an external indel with not enough read support
-    const bool isCurIndelNoReadSupport(curIndelData.status.noReadSupport);
+    // true if current indel is an external indel not discovered from reads
+    const bool isCurIndelNotDiscoveredFromReads(curIndelData.status.notDiscoveredFromReads);
 
     // Get haplotype IDs of current indel in all samples
     std::vector<int> curIndelHaplotypeIds(sampleCount);
@@ -1034,47 +1038,44 @@ candidate_alignment_search(
     // alignment 1) --> unchanged case:
     try
     {
-        bool isValid(true);
+        bool isNextCandidateAlignmentValid(true);
 
-        // Check haplotype constraints
+        // 1. Check haplotype constraints
         HaplotypeStatusMap newHaplotypeStatusMap(haplotypeStatusMap);
         if ((! isCurIndelConflicting) && isCurIndelInActiveRegion)
         {
-            isValid = newHaplotypeStatusMap.at(curIndelActiveRegionId).updateHaplotypeStatus(curIndelHaplotypeIds, isCurIndelOn);
+            isNextCandidateAlignmentValid = newHaplotypeStatusMap.at(curIndelActiveRegionId)
+                .updateHaplotypeStatus(curIndelHaplotypeIds, isCurIndelOn);
         }
         else
         {
             // current indel is conflicting or there's no haplotype info
-            isValid = (! curIndel.isMismatch()) || (! isCurIndelOn);
+            isNextCandidateAlignmentValid = (! curIndel.isMismatch()) || (! isCurIndelOn);
         }
 
-        // Check whether we will be including
-        // more than one indels with no read support
-        bool newContainsIndelWithNoReadSupport(containsIndelWithNoReadSupport);
-        if (isCurIndelOn && isCurIndelNoReadSupport)
+        // 2. Check whether we will be including
+        // more than one indels without enough read support
+        if (isCurIndelOn && containsIndelNotDiscoveredFromReads &&
+            isCurIndelNotDiscoveredFromReads)
         {
-            if (containsIndelWithNoReadSupport)
-            {
-                // it's prohibited to include more than one indels with no read support
-                isValid = false;
-            }
-            newContainsIndelWithNoReadSupport = true;
+            // it's prohibited to include more than one indels without enough read support
+            isNextCandidateAlignmentValid = false;
         }
 
-        if ((! isValid) && (totalToggleDepth == 0))
+        if ((! isNextCandidateAlignmentValid) && (totalToggleDepth == 0))
         {
-            // even if the above constraints are not met,
+            // even if the above conditions are not met,
             // if there was no indel toggle,
             // allow the alignment search to proceed
-            isValid = true;
+            isNextCandidateAlignmentValid = true;
         }
 
-        if (isValid)
+        if (isNextCandidateAlignmentValid)
         {
             candidate_alignment_search(opt, dopt, read_id, read_length, indelBuffer,
                                        sampleId, realign_buffer_range,
                                        cal_set, warn,
-                                       indel_status_map, newHaplotypeStatusMap, newContainsIndelWithNoReadSupport,
+                                       indel_status_map, newHaplotypeStatusMap,
                                        indel_order, depth + 1, indelToggleDepth, totalToggleDepth,
                                        read_range, max_read_indel_toggle, cal);
         }
@@ -1087,31 +1088,28 @@ candidate_alignment_search(
         throw;
     }
 
-    bool isValid(true);
+    bool isNextCandidateAlignmentValid(true);
     HaplotypeStatusMap newHaplotypeStatusMap(haplotypeStatusMap);
     if (!isCurIndelConflicting && isCurIndelInActiveRegion)
     {
-        isValid = newHaplotypeStatusMap.at(curIndelActiveRegionId).updateHaplotypeStatus(curIndelHaplotypeIds, !isCurIndelOn);
+        isNextCandidateAlignmentValid = newHaplotypeStatusMap.at(curIndelActiveRegionId).updateHaplotypeStatus(curIndelHaplotypeIds, !isCurIndelOn);
     }
     else
     {
-        isValid = !curIndel.isMismatch() || isCurIndelOn;
+        isNextCandidateAlignmentValid = !curIndel.isMismatch() || isCurIndelOn;
     }
 
     // Check whether we will be including
-    // more than one indels with no read support
-    bool newContainsIndelWithNoReadSupport(containsIndelWithNoReadSupport);
-    if (!isCurIndelOn && isCurIndelNoReadSupport)
+    // more than one indels not discovered from reads
+    if (!isCurIndelOn &&
+        containsIndelNotDiscoveredFromReads &&
+        isCurIndelNotDiscoveredFromReads)
     {
-        if (containsIndelWithNoReadSupport)
-        {
-            // it's prohibited to include more than one indels with no read support
-            isValid = false;
-        }
-        newContainsIndelWithNoReadSupport = true;
+        // it's prohibited to include more than one indels without enough read support
+        isNextCandidateAlignmentValid = false;
     }
 
-    if (! isValid) return;
+    if (! isNextCandidateAlignmentValid) return;
 
     if (! isCurIndelOn)
     {
@@ -1193,7 +1191,7 @@ candidate_alignment_search(
                 candidate_alignment_search(opt, dopt, read_id, read_length, indelBuffer,
                                            sampleId,
                                            realign_buffer_range, cal_set,
-                                           warn, indel_status_map, newHaplotypeStatusMap, newContainsIndelWithNoReadSupport,
+                                           warn, indel_status_map, newHaplotypeStatusMap,
                                            indel_order, depth + 1, indelToggleDepth + indelToggleIncrement, totalToggleDepth + 1,
                                            read_range, max_read_indel_toggle,
                                            start_cal);
@@ -1262,7 +1260,7 @@ candidate_alignment_search(
 
                     candidate_alignment_search(opt, dopt, read_id, read_length, indelBuffer, sampleId,
                                                realign_buffer_range, cal_set,
-                                               warn, indel_status_map, newHaplotypeStatusMap, newContainsIndelWithNoReadSupport,
+                                               warn, indel_status_map, newHaplotypeStatusMap,
                                                indel_order, depth + 1, indelToggleDepth + indelToggleIncrement, totalToggleDepth + 1,
                                                read_range, max_read_indel_toggle, start_cal);
                 }
@@ -1954,11 +1952,10 @@ getCandidateAlignments(
     static const unsigned startDepth(0);
     static const unsigned startIndelToggleDepth(0);
     static const unsigned startTotalToggleDepth(0);
-    static const bool startContainsIndelWithNoReadSupport(false);
 
     candidate_alignment_search(opt, dopt, rseg.getReadIndex(), cal_read_length, indelBuffer,
                                sampleId, realign_buffer_range, cal_set,
-                               warn, indel_status_map, haplotypeStatusMap, startContainsIndelWithNoReadSupport,
+                               warn, indel_status_map, haplotypeStatusMap,
                                indel_order, startDepth, startIndelToggleDepth, startTotalToggleDepth,
                                exemplar_pr, opt.max_read_indel_toggle, cal);